+++ buildchain/buildchain/image.py
@@ -268,7 +268,7 @@ TO_BUILD: Tuple[targets.LocalImage, ...] = (
         name="salt-master",
         build_args={
             "BASE_IMAGE": versions.ROCKY_BASE_IMAGE,
-            "BASE_IMAGE_SHA256": versions.ROCKY_BASE_IMAGE_8_SHA256,
+            "BASE_IMAGE_SHA256": versions.ROCKY_BASE_IMAGE_9_SHA256,
             "SALT_VERSION": versions.SALT_VERSION,
         },
     ),