+++ buildchain/buildchain/targets/package.py
@@ -9,7 +9,7 @@ This modules provides several services:
 - download the source files
 - build a SRPM from the source files
 
-Note that for now, it only works for CentOS 7 x86_64.
+Note that for now, it only works for Rocky/RedHat 8 and 9 x86_64.
 
             self.make_package_directory(),
             self.generate_meta(),
@@ -82,7 +82,7 @@ class Package(base.CompositeTarget):
 
 
 class RPMPackage(Package):
-    """A RPM software package for CentOS."""
+    """A RPM software package for Rocky/RedHat."""
 
     SOURCE_URL_PATTERN = re.compile(r"^(Source|Patch)\d+:\s+(?P<url>.+)$")